+++ src/Data/decklists/Luxury Riftbound Champion Series.ts
@@ -1,0 +1,28 @@
+import { PATCH } from "../Enums";
+import { TournamentResults } from "../Interfaces";
+
+export const LRCS_20250816: TournamentResults = {
+    tournamentName: "Luxury Riftbound Champion Series", abbrName: "LRCS", date: "2025/08/16", meta: PATCH.OGN, host: "Luxury Gaming",
+    links: ["https://luxurygaming.challonge.com/LRCS#"],
+    placings: [
+        {placing: "6-0", decklists: [
+            {"username":"tog","date":"2025/08/16","archetype":"Sett Midrange","legend":"OGN-269-1","chosenChampion":"OGN-240-1","mainDeck":[{"id":"OGN-240-1","count":1},{"id":"OGN-164-2","count":3},{"id":"OGN-270-1","count":3},{"id":"OGN-156-1","count":3},{"id":"OGN-136-1","count":3},{"id":"OGN-139-1","count":3},{"id":"OGN-213-1","count":2},{"id":"OGN-210-1","count":3},{"id":"OGN-128-1","count":2},{"id":"OGN-217-1","count":3},{"id":"OGN-132-1","count":3},{"id":"OGN-207-1","count":3},{"id":"OGN-141-1","count":3},{"id":"OGN-155-1","count":3},{"id":"OGN-149-1","count":2}],"battlefields":[{"id":"OGN-297-1","count":1},{"id":"OGN-282-1","count":1},{"id":"OGN-284-1","count":1}],"runeDeck":[{"id":"OGN-214-2","count":5},{"id":"OGN-126-2","count":7}],"sideboard":[{"id":"OGN-213-1","count":1},{"id":"OGN-238-1","count":3},{"id":"OGN-232-1","count":3},{"id":"OGN-149-1","count":1}]},
+        ]},
+        {placing: "5-0-1", decklists: [
+            {"username":"WCC_Kendall","date":"2025/08/16","archetype":"Master Yi Midrange","legend":"OGS-019-1","chosenChampion":"OGS-009-1","mainDeck":[{"id":"OGS-009-1","count":1},{"id":"OGN-045-1","count":3},{"id":"OGN-046-1","count":3},{"id":"OGN-043-1","count":3},{"id":"OGN-136-1","count":3},{"id":"OGN-052-1","count":3},{"id":"OGN-058-1","count":3},{"id":"OGN-044-1","count":3},{"id":"OGN-077-1","count":3},{"id":"OGN-054-1","count":2},{"id":"OGN-132-1","count":3},{"id":"OGN-154-1","count":2},{"id":"OGN-155-1","count":3},{"id":"OGN-149-1","count":3},{"id":"OGN-082-1","count":2}],"battlefields":[{"id":"OGN-280-1","count":1},{"id":"OGN-292-1","count":1},{"id":"OGN-288-1","count":1}],"runeDeck":[{"id":"OGN-042-1","count":6},{"id":"OGN-126-1","count":6}],"sideboard":[{"id":"OGN-145-1","count":3},{"id":"OGN-063-1","count":3},{"id":"OGN-164-1","count":2}]},
+            {"username":"Robomaster18","date":"2025/08/16","archetype":"","legend":"","chosenChampion":"","mainDeck":[],"battlefields":[],"runeDeck":[],"sideboard":[]},
+        ]},
+        {placing: "5-1", decklists: [
+            {"username":"Prismaticism","date":"2025/08/16","archetype":"","legend":"","chosenChampion":"","mainDeck":[],"battlefields":[],"runeDeck":[],"sideboard":[]},
+            {"username":"LNDfr3nk","date":"2025/08/16","archetype":"","legend":"","chosenChampion":"","mainDeck":[],"battlefields":[],"runeDeck":[],"sideboard":[]},
+            {"username":"Batatanator","date":"2025/08/16","archetype":"","legend":"","chosenChampion":"","mainDeck":[],"battlefields":[],"runeDeck":[],"sideboard":[]},
+            {"username":"Kuvi","date":"2025/08/16","archetype":"","legend":"","chosenChampion":"","mainDeck":[],"battlefields":[],"runeDeck":[],"sideboard":[]},
+            {"username":"ManuS","date":"2025/08/16","archetype":"","legend":"","chosenChampion":"","mainDeck":[],"battlefields":[],"runeDeck":[],"sideboard":[]},
+            {"username":"powerhous230","date":"2025/08/16","archetype":"","legend":"","chosenChampion":"","mainDeck":[],"battlefields":[],"runeDeck":[],"sideboard":[]},
+            {"username":"Vendorty","date":"2025/08/16","archetype":"","legend":"","chosenChampion":"","mainDeck":[],"battlefields":[],"runeDeck":[],"sideboard":[]},
+            {"username":"ConnectedTwig","date":"2025/08/16","archetype":"","legend":"","chosenChampion":"","mainDeck":[],"battlefields":[],"runeDeck":[],"sideboard":[]},
+            {"username":"Yooric","date":"2025/08/16","archetype":"","legend":"","chosenChampion":"","mainDeck":[],"battlefields":[],"runeDeck":[],"sideboard":[]},
+            {"username":"deluxephilcheeze","date":"2025/08/16","archetype":"","legend":"","chosenChampion":"","mainDeck":[],"battlefields":[],"runeDeck":[],"sideboard":[]},
+        ]}
+    ]
+};