+++ src/Pages/index.ts
@@ -1,5 +1,7 @@
+export * from "./AllArchetypesPage";
 export * from "./CardListPage";
 export * from "./DecklistTablePage";
 export * from "./ErrorPage";
 export * from "./GenerateDecklistPage";
+export * from "./PlayerResultsPage";
 export * from "./TournamentResultsPage";